[project-management-system#ai-service]: Moved JwtVerificationTokenFilter filter to seperated folder
